==> code/01_Concatenating_clean_train_test_csvs.ipynb <==
# Combining Train and Test Data from Separate EDA Processes

Notebook to concatenate separetly cleaned sections of the train and test data sets.

Magnus Bigelow

## Contents

- [Training Data](#Training-Data)
- [Testing Data](#Testing-Data)

import pandas as pd

## Training Data

train_1 = pd.read_csv('../data/clean_train_1.csv')
train_2 = pd.read_csv('../data/clean_train_2.csv')

train_1.head()

train_2.head()

# concatenate
train_clean = pd.concat([train_1,train_2], axis=1)

train_clean.head()

# Save to csv
train_clean.to_csv('../data/clean_train.csv', index=False)

train_clean.shape

train_1.shape

train_2.shape

## Testing Data

test_1 = pd.read_csv('../data/clean_test_1.csv')
test_2 = pd.read_csv('../data/clean_test_2.csv')

test_1.shape

test_2.shape

Shapes look good

# concatenate
test_clean = pd.concat([test_1,test_2], axis=1)

test_clean.head()

# Save to csv
test_clean.to_csv('../data/clean_test.csv', index=False)
